refactored code for SRP, DRY
