+++ LoginManager/SessionHandler/ISession.cs
