+++ LoginManager/SessionHandler/Session.cs
