+++ LoginManager/SessionHandler/SessionContext.cs
@@ -1,0 +1,6 @@
+namespace SessionHandler;
+
+public class SessionContext
+{
+    
+}
